Extract out useEntity into api folder & use useEntity instead of dedicated fetch endpoints for each entity - the latter just reduces some repeated code
