+++ packages/datagateway-common/src/api/index.tsx
@@ -1,38 +1,39 @@
-import React from 'react';
+import {
+  UseQueryOptions,
+  UseQueryResult,
+  useQueries,
+  useQuery,
+} from '@tanstack/react-query';
 import axios, { AxiosError } from 'axios';
+import { isValid } from 'date-fns';
+import format from 'date-fns/format';
+import React from 'react';
+import { useSelector } from 'react-redux';
 import { useHistory, useLocation } from 'react-router-dom';
 import {
   AdditionalFilters,
-  FiltersType,
-  SortType,
-  Order,
-  Filter,
-  QueryParams,
-  ViewsType,
   Entity,
+  Filter,
+  FiltersType,
+  Order,
+  QueryParams,
+  SortType,
   UpdateMethod,
+  ViewsType,
 } from '../app.types';
-import {
-  useQueries,
-  useQuery,
-  UseQueryOptions,
-  UseQueryResult,
-} from '@tanstack/react-query';
 import handleICATError from '../handleICATError';
 import { readSciGatewayToken } from '../parseTokens';
-import { useSelector } from 'react-redux';
 import { StateType } from '../state/app.types';
-import format from 'date-fns/format';
-import { isValid } from 'date-fns';
 import { useRetryICATErrors } from './retryICATErrors';
 
 export * from './cart';
+export * from './dataPublications';
+export * from './datafiles';
+export * from './datasets';
 export * from './facilityCycles';
+export * from './generic';
 export * from './instruments';
 export * from './investigations';
-export * from './datafiles';
-export * from './dataPublications';
-export * from './datasets';
 export * from './lucene';
 
 /**
